Bugfix/arduino syntax fix

+++ src/miniauto/libraries/Melopero_AMG8833-1.1.0/src/Melopero_AMG8833.h
@@ -68,7 +68,7 @@ class Melopero_AMG8833 {
         uint8_t i2cAddress;
         uint8_t bus;
 
-        float pixelMatrix[8][8];
+        int16_t pixelMatrix[8][8];
         float thermistorTemperature;
 
         bool interruptMatrix[8][8];
